requirements.txt
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dj-rest-auth==7.0.1
Django==5.2b1
django-activity-stream==2.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-js-asset==3.0.1
django-mptt==0.16.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
furl==2.1.3
gunicorn==23.0.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
orderedmultidict==1.0.1
packaging==24.2
prompt_toolkit==3.0.50
psycopg2==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
RapidFuzz==3.12.1
redis==5.2.1
referencing==0.36.1
rpds-py==0.22.3
setuptools==75.8.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
